Update dependencies and refactor project structure
